kb/scraping/sentenze-cassazione/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-storage==2.14.0
google-auth==2.25.2
python-dotenv==1.0.0
